+++ piecrust/src/imports.rs
@@ -182,7 +182,7 @@ fn c(
 
     let ret = match instance.with_memory_mut(with_memory) {
         Ok((ret_len, callee_spent)) => {
-            env.move_up_call_tree();
+            env.move_up_call_tree(callee_spent);
             instance.set_remaining_points(caller_remaining - callee_spent);
             ret_len
         }